добавил интерфейс ObjectMapper и отредактировал ExceptionHandlerAdvice
